added EmployeesGrossSalary collection to firebase fixed bug in AddUserFragment that don't save phone number in the firebase fixed bug in UserFragmentDialog that when display information the phone number and password don't show up

+++ IGEC-User/app/src/main/java/com/example/igecuser/fireBase/Allowance.java
@@ -3,8 +3,8 @@ package com.example.igecuser.fireBase;
 import java.io.Serializable;
 
 public class Allowance implements Serializable {
-    private String name ;
-    private int amount ;
+    private String name;
+    private double amount;
 
     public Allowance() {
     }
@@ -17,15 +17,15 @@ public class Allowance implements Serializable {
         this.name = name;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
 
-    public Allowance(String name, int amount) {
+    public Allowance(String name, double amount) {
         this.name = name;
         this.amount = amount;
     }
